Rename input to text_input and update references and test filenames
Fix failing dictionary combination syntax
